+++ seadex_sonarr/modules/torrent.py
@@ -1,0 +1,14 @@
+import pynyaa
+
+
+def get_nyaa_url(url):
+    """Get Nyaa torrent link from URL
+
+    Args:
+        url (str): URL to get Nyaa torrent link
+    """
+
+    nyaa = pynyaa.get(url)
+    parsed_url = str(nyaa.torrent_file)
+
+    return parsed_url