+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.0.2 (2025-07-13)
 ==================
 
+- Add interactive mode, for selecting when multiple "best" options are found
+- Add support for adding torrents to qBittorrent
 - Improved Discord messaging
 - Catch the case where we don't find any suitable SeaDex releases
 - Include potentially weird offset mappings via AniDB lists